CM11 Themes: Store hash codes for target and overlay in idmap
Due to some system packages always having the same last modified
time stamp of 08/01/2008 we must use some other unique value to
determine that the contents of either the target or the overlay
have changed.  This unique value will be passed down from the java
layer and written to the idmap header.

This change requires the following patch from frameworks/base
http://review.cyanogenmod.org/62565

Change-Id: I0adcafaca96110268766d6e3349414364ecea4c3

Conflicts:
	cmds/installd/commands.c
	cmds/installd/installd.c
	cmds/installd/installd.h